requirements.txt
requests~=2.31.0
Django~=4.2.2
djangorestframework~=3.14.0
mail-parser==3.15.0
django-dotenv==1.4.2